pytorch model

+++ Python3-MODELS/Juan_Esteban/Pytorch_model/Inversion/stokes_as_channels/muram.py
@@ -19,8 +19,13 @@ from tqdm import tqdm
 
 class MuRAM():
     #To rescale all the data e are going to use a max values of the order of the general maximum value of the data, and 
-    def __init__(self, ptm, filename):
+    def __init__(self, ptm:str, pth_out:str, filename:str):
+        """
+        ptm (str): Path for MURAM data.
+        pth_out (str): Path for output data.
+        """
         self.ptm = ptm
+        self.pth_out = pth_out
         self.filename = filename
         self.nlam = 300 #this parameter is useful when managing the Stokes parameters #wavelenght interval - its from 6300 amstroengs in steps of 10 amstroengs
         self.nx = 480
@@ -111,14 +116,15 @@ class MuRAM():
             print("Applying optical depth stratification...")
             opt_depth = np.load(self.ptm+"optical_depth_"+self.filename+".npy")
             #optical depth points
-            tau_out = "self.ptm+array_of_tau_"+self.filename+f"_{opt_len}_depth_points.npy"
-            tau = np.linspace(-3, 1, opt_len)
-            np.save(tau_out, tau)
+            tau_out = self.pth_out+"array_of_tau_"+self.filename+f"_{opt_len}_depth_points.npy"
+            if not os.path.exists(tau_out):
+                tau = np.linspace(-3, 1, opt_len)
+                np.save(tau_out, tau)
 
             #optical stratification
             opt_mags_interp = {}
             opt_mags = np.zeros((self.nx, opt_len, self.nz, atm_quant.shape[-1]))
-            opt_mags_out ="self.ptm+optical_stratified_atm_"+self.filename+f"_{opt_len}_depth_points.npy"
+            opt_mags_out =self.pth_out+"optical_stratified_atm_"+self.filename+f"_{opt_len}_depth_points.npy"
             if not os.path.exists(opt_mags_out):
                 for ix in tqdm(range(self.nx)):
                         for iz in range(self.nz):
@@ -246,10 +252,11 @@ class MuRAM():
     def train_test_sets(self, name_of_input, gran_inter_zones = False, scale = True, opt_depth_stratif = True, opt_len = 20, vertical_comp = True):
         atm_quant, stokes = self.granular_intergranular(gran_inter_zones = gran_inter_zones, scale = scale, opt_depth_stratif = opt_depth_stratif, opt_len = opt_len, vertical_comp = vertical_comp)
         print("splitting...")
+        random_seed = 42
         if name_of_input == "Stokes":
-            in_train, in_test, out_train, out_test = train_test_split(stokes, atm_quant, test_size=0.33, random_state=42)
+            in_train, in_test, out_train, out_test = train_test_split(stokes, atm_quant, test_size=0.33, random_state=random_seed)
         elif name_of_input == "Atm":
-            in_train, in_test, out_train, out_test = train_test_split(stokes, atm_quant, test_size=0.33, random_state=42)
+            in_train, in_test, out_train, out_test = train_test_split(stokes, atm_quant, test_size=0.33, random_state=random_seed)
         else:
             raise ValueError("Not possible input")
